FU-104 refactor: ProfileService 인터페이스를 삭제 후 구현체로만 관리
- 아직 인터페이스의 분리에 대한 안목이 보이지 않는 시점이기에 오버 엔지니어링을 방지하였음
